+++ src/core/components/accordion/accordionContainer/index.ts
@@ -1,0 +1,1 @@
+export { AccordionContainer, type IAccordionContainerProps } from './accordionContainer';