+++ src/core/components/accordion/accordionItemHeader/index.ts
@@ -1,0 +1,1 @@
+export { AccordionItemHeader } from './accordionItemHeader';