+++ CHANGELOG.md
@@ -9,7 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ### Added
 
--   Implement `Accordion` core components
+-   Implement `Accordion` core component
 -   Implement `ProposalDataListItemSkeleton` module component
 
 ## [1.0.23] - 2024-04-18